backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
psycopg2-binary==2.9.9
redis==5.0.1
python-dotenv==1.0.0
Pillow==10.1.0
django-filter==23.5
drf-yasg==1.21.7
setuptools<81
celery==5.3.4
django-redis==5.4.0
channels==4.0.0
channels-redis==4.1.0
gunicorn==21.2.0
whitenoise==6.6.0

